Refactor user registration logic into authController and update usersRouter to use the new controller method
